extract IntUiBridgeRadioButton
