+++ ide-laf-bridge/src/main/kotlin/org/jetbrains/jewel/bridge/theme/IntUiBridge.kt
@@ -24,8 +24,6 @@ import com.intellij.ui.JBColor
 import com.intellij.util.ui.DirProvider
 import com.intellij.util.ui.JBUI
 import com.intellij.util.ui.NamedColorUtil
-import javax.swing.UIManager
-import kotlin.time.Duration.Companion.milliseconds
 import org.jetbrains.jewel.bridge.createVerticalBrush
 import org.jetbrains.jewel.bridge.dp
 import org.jetbrains.jewel.bridge.isNewUiTheme
@@ -95,10 +93,6 @@ import org.jetbrains.jewel.ui.component.styling.MenuStyle
 import org.jetbrains.jewel.ui.component.styling.PopupContainerColors
 import org.jetbrains.jewel.ui.component.styling.PopupContainerMetrics
 import org.jetbrains.jewel.ui.component.styling.PopupContainerStyle
-import org.jetbrains.jewel.ui.component.styling.RadioButtonColors
-import org.jetbrains.jewel.ui.component.styling.RadioButtonIcons
-import org.jetbrains.jewel.ui.component.styling.RadioButtonMetrics
-import org.jetbrains.jewel.ui.component.styling.RadioButtonStyle
 import org.jetbrains.jewel.ui.component.styling.SegmentedControlButtonColors
 import org.jetbrains.jewel.ui.component.styling.SegmentedControlButtonMetrics
 import org.jetbrains.jewel.ui.component.styling.SegmentedControlButtonStyle
@@ -127,8 +121,9 @@ import org.jetbrains.jewel.ui.component.styling.TextFieldStyle
 import org.jetbrains.jewel.ui.component.styling.TooltipColors
 import org.jetbrains.jewel.ui.component.styling.TooltipMetrics
 import org.jetbrains.jewel.ui.component.styling.TooltipStyle
-import org.jetbrains.jewel.ui.icon.PathIconKey
 import org.jetbrains.jewel.ui.icons.AllIconsKeys
+import javax.swing.UIManager
+import kotlin.time.Duration.Companion.milliseconds
 
 private val logger = JewelLogger.getInstance("JewelIntUiBridge")
 
@@ -755,79 +750,6 @@ private fun readMenuStyle(): MenuStyle {
             ),
         icons = MenuIcons(submenuChevron = AllIconsKeys.General.ChevronRight),
     )
-}
-
-private fun readRadioButtonStyle(): RadioButtonStyle {
-    val normalContent = retrieveColorOrUnspecified("RadioButton.foreground")
-    val disabledContent = retrieveColorOrUnspecified("RadioButton.disabledText")
-    val colors =
-        RadioButtonColors(
-            content = normalContent,
-            contentHovered = normalContent,
-            contentDisabled = disabledContent,
-            contentSelected = normalContent,
-            contentSelectedHovered = normalContent,
-            contentSelectedDisabled = disabledContent,
-        )
-
-    val newUiTheme = isNewUiTheme()
-    val metrics = if (newUiTheme) NewUiRadioButtonMetrics else ClassicUiRadioButtonMetrics
-
-    // This value is not normally defined in the themes, but Swing checks it anyway
-    // The default hardcoded in
-    // com.intellij.ide.ui.laf.darcula.ui.DarculaRadioButtonUI.getDefaultIcon()
-    // is not correct though, the SVG is 19x19 and is missing 1px on the right
-    val radioButtonSize =
-        retrieveIntAsDpOrUnspecified("RadioButton.iconSize")
-            .takeOrElse { metrics.radioButtonSize }
-            .let { DpSize(it, it) }
-
-    // val outlineSize = if (isNewUiButNotDarcula() DpSize(17.dp, 17.dp) else
-
-    return RadioButtonStyle(
-        colors = colors,
-        metrics =
-            RadioButtonMetrics(
-                radioButtonSize = radioButtonSize,
-                outlineSize = metrics.outlineSize,
-                outlineFocusedSize = metrics.outlineFocusedSize,
-                outlineSelectedSize = metrics.outlineSelectedSize,
-                outlineSelectedFocusedSize = metrics.outlineSelectedFocusedSize,
-                iconContentGap =
-                    retrieveIntAsDpOrUnspecified("RadioButton.textIconGap").takeOrElse { metrics.iconContentGap },
-            ),
-        icons = RadioButtonIcons(radioButton = PathIconKey("${iconsBasePath}radio.svg", RadioButtonIcons::class.java)),
-    )
-}
-
-private interface BridgeRadioButtonMetrics {
-    val outlineSize: DpSize
-    val outlineFocusedSize: DpSize
-    val outlineSelectedSize: DpSize
-    val outlineSelectedFocusedSize: DpSize
-
-    val radioButtonSize: Dp
-    val iconContentGap: Dp
-}
-
-private object ClassicUiRadioButtonMetrics : BridgeRadioButtonMetrics {
-    override val outlineSize = DpSize(17.dp, 17.dp)
-    override val outlineFocusedSize = DpSize(19.dp, 19.dp)
-    override val outlineSelectedSize = outlineSize
-    override val outlineSelectedFocusedSize = outlineFocusedSize
-
-    override val radioButtonSize = 19.dp
-    override val iconContentGap = 4.dp
-}
-
-private object NewUiRadioButtonMetrics : BridgeRadioButtonMetrics {
-    override val outlineSize = DpSize(17.dp, 17.dp)
-    override val outlineFocusedSize = outlineSize
-    override val outlineSelectedSize = DpSize(22.dp, 22.dp)
-    override val outlineSelectedFocusedSize = outlineSelectedSize
-
-    override val radioButtonSize = 24.dp
-    override val iconContentGap = 4.dp
 }
 
 private fun readSegmentedControlButtonStyle(): SegmentedControlButtonStyle {
